pyproject.toml
[project]
name = "predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]
